# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Wildflower Fundraising CRM API (Airtable-aligned schema)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /media-mentions:
    get:
      operationId: listMediaMentions
      tags: [media-mentions]
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: personId, in: query, schema: { type: string } }
        - { name: organizationId, in: query, schema: { type: string } }
        - { name: pinned, in: query, schema: { type: boolean }, description: "Filter to pinned (true) or unpinned (false) mentions." }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Page" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/MediaMentionList" } } } }
    post:
      operationId: createMediaMention
      tags: [media-mentions]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/MediaMentionInput" } } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/MediaMention" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /media-mentions/{id}:
    get:
      operationId: getMediaMention
      tags: [media-mentions]
      parameters: [ { $ref: "#/components/parameters/IdPath" } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/MediaMention" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateMediaMention
      tags: [media-mentions]
      parameters: [ { $ref: "#/components/parameters/IdPath" } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/MediaMentionUpdate" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/MediaMention" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteMediaMention
      tags: [media-mentions]
      parameters: [ { $ref: "#/components/parameters/IdPath" } ]
      responses: { "204": { description: Deleted } }

  # ─── Tasks ────────────────────────────────────────────────────────────────
...
  /meeting-notes:
    get:
      operationId: listMeetingNotes
      tags: [meeting-notes]
      parameters:
        - { name: personId, in: query, schema: { type: string } }
        - { name: organizationId, in: query, schema: { type: string } }
        - { name: householdId, in: query, schema: { type: string } }
        - { name: creatorUserId, in: query, schema: { type: string } }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Page" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/MeetingNoteList" } } } }
    post:
      operationId: createMeetingNote
      tags: [meeting-notes]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateMeetingNoteBody" } } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/MeetingNote" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /meeting-notes/{id}:
    get:
      operationId: getMeetingNote
      tags: [meeting-notes]
      parameters: [ { $ref: "#/components/parameters/IdPath" } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/MeetingNote" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateMeetingNote
      tags: [meeting-notes]
      parameters: [ { $ref: "#/components/parameters/IdPath" } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/UpdateMeetingNoteBody" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/MeetingNote" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteMeetingNote
      tags: [meeting-notes]
      parameters: [ { $ref: "#/components/parameters/IdPath" } ]
      responses: { "204": { description: Deleted } }
  /meeting-notes/{id}/promote-action-item:
    post:
      operationId: promoteMeetingActionItem
      tags: [meeting-notes]
      parameters: [ { $ref: "#/components/parameters/IdPath" } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/PromoteActionItemBody" } } } }
      responses:
        "201": { description: Task created from action item, content: { application/json: { schema: { $ref: "#/components/schemas/Task" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── Saved views (per-list-page filter+sort presets) ───────────────────────
...
components:
  parameters:
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 10000, default: 50 } }
    Page:  { name: page,  in: query, schema: { type: integer, minimum: 1, default: 1 } }
    IdPath: { name: id, in: path, required: true, schema: { type: string } }
  responses:
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
    BadRequest:
      description: Bad request (validation error)
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:   { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
    Pagination:
      type: object
      required: [page, limit, total]
      properties:
        page:  { type: integer }
        limit: { type: integer }
        total: { type: integer }

    # ─── Enums ─────────────────────────────────────────────────────────────────
...
    MediaMention:
      type: object
      required: [id, publicationName, url, pinned, createdAt, updatedAt]
      properties:
        id:              { type: string }
        publicationName: { type: string }
        title:           { type: string, nullable: true }
        author:          { type: string, nullable: true }
        publicationDate: { type: string, format: date, nullable: true }
        url:             { type: string }
        aiSummary:       { type: string, nullable: true }
        source:          { type: string, nullable: true }
        pinned:          { type: boolean }
        personIds:       { type: array, items: { type: string }, nullable: true }
        organizationIds:       { type: array, items: { type: string }, nullable: true }
        createdAt:       { type: string, format: date-time }
        updatedAt:       { type: string, format: date-time }
    MediaMentionList:
      type: object
      required: [data, pagination]
      properties:
        data:       { type: array, items: { $ref: "#/components/schemas/MediaMention" } }
        pagination: { $ref: "#/components/schemas/Pagination" }
    MediaMentionInput:
      type: object
      required: [publicationName, url]
      properties:
        publicationName: { type: string, minLength: 1 }
        title:           { type: string }
        author:          { type: string }
        publicationDate: { type: string, format: date }
        url:             { type: string, minLength: 1 }
        aiSummary:       { type: string }
        source:          { type: string }
        pinned:          { type: boolean }
        personIds:       { type: array, items: { type: string } }
        organizationIds:       { type: array, items: { type: string } }
    MediaMentionUpdate:
      type: object
      properties:
        publicationName: { type: string, minLength: 1 }
        title:           { type: string, nullable: true }
        author:          { type: string, nullable: true }
        publicationDate: { type: string, format: date, nullable: true }
        url:             { type: string, minLength: 1 }
        aiSummary:       { type: string, nullable: true }
        source:          { type: string, nullable: true }
        pinned:          { type: boolean }
        personIds:       { type: array, items: { type: string }, nullable: true }
        organizationIds:       { type: array, items: { type: string }, nullable: true }

    # ─── Tasks ───────────────────────────────────────────────────────────────
    TaskStatus:
      type: string
      enum: [open, waiting, done, cancelled]
    TaskKind:
      type: string
      enum: [general, reporting_deadline, thank_you_followup]
    Task:
      type: object
      required: [id, title, status, kind, createdByUserId, createdAt, updatedAt]
      properties:
        id:               { type: string }
        title:            { type: string }
        description:      { type: string, nullable: true }
        dueDate:          { type: string, format: date, nullable: true }
        kind:             { $ref: "#/components/schemas/TaskKind" }
        status:           { $ref: "#/components/schemas/TaskStatus" }
        completedAt:      { type: string, format: date-time, nullable: true }
        assigneeUserId:   { type: string, nullable: true }
        createdByUserId:  { type: string }
        personIds:        { type: array, items: { type: string }, nullable: true }
        organizationIds:        { type: array, items: { type: string }, nullable: true }
        householdIds:     { type: array, items: { type: string }, nullable: true }
        opportunityIds:   { type: array, items: { type: string }, nullable: true }
        giftIds:          { type: array, items: { type: string }, nullable: true }
        mentionUserIds:   { type: array, items: { type: string }, nullable: true }
        createdAt:        { type: string, format: date-time }
        updatedAt:        { type: string, format: date-time }
...
    MeetingActionItem:
      type: object
      required: [title]
      properties:
        title:           { type: string }
        assigneeName:    { type: string, nullable: true, description: "Free-text assignee name as extracted from the transcript." }
        dueDate:         { type: string, format: date, nullable: true }
        promotedTaskId:  { type: string, nullable: true, description: "Set after this item was promoted into a task via /promote-action-item." }
    MeetingNote:
      type: object
      required: [id, meetingDate, summaryOnly, creatorUserId, createdAt, updatedAt]
      properties:
        id:               { type: string }
        title:            { type: string, nullable: true }
        meetingDate:      { type: string, format: date-time }
        attendees:        { type: array, items: { type: string }, nullable: true }
        rawTranscript:    { type: string, nullable: true, description: "Always null when summaryOnly is true." }
        summaryOnly:      { type: boolean, description: "Snapshot of the creator's email_sync_mode at create time. When true, the raw transcript was dropped pre-insert." }
        aiSummary:        { type: string, nullable: true }
        actionItems:
          type: array
          nullable: true
          items: { $ref: "#/components/schemas/MeetingActionItem" }
        creatorUserId:    { type: string }
        personId:         { type: string, nullable: true }
        organizationId:         { type: string, nullable: true }
        householdId:      { type: string, nullable: true }
        createdAt:        { type: string, format: date-time }
        updatedAt:        { type: string, format: date-time }
    MeetingNoteList:
      type: object
      required: [data, pagination]
      properties:
        data:       { type: array, items: { $ref: "#/components/schemas/MeetingNote" } }
        pagination: { $ref: "#/components/schemas/Pagination" }
    CreateMeetingNoteBody:
      type: object
      description: "Exactly one of personId / funderId / householdId must be set (contact XOR). Exactly one of `transcript` or `summary` must be provided — `transcript` runs through AI summarization, `summary` is stored verbatim as the note body (used by the hand-typed-notes flow)."
      properties:
        transcript:      { type: string, description: "Raw pasted transcript. Dropped server-side before insert when the caller's email_sync_mode is summary_only. Runs through AI summarization to produce aiSummary + actionItems." }
        summary:         { type: string, description: "Hand-typed notes. Stored verbatim as aiSummary with no AI processing; no rawTranscript or actionItems are generated. Mutually exclusive with transcript." }
        title:           { type: string }
        meetingDate:     { type: string, format: date-time, description: "Defaults to now if omitted." }
        attendees:       { type: array, items: { type: string } }
        personId:        { type: string }
        organizationId:        { type: string }
        householdId:     { type: string }
    UpdateMeetingNoteBody:
      type: object
      properties:
        title:           { type: string, nullable: true }
        meetingDate:     { type: string, format: date-time }
        attendees:       { type: array, items: { type: string }, nullable: true }
        aiSummary:       { type: string, nullable: true }
        actionItems:
          type: array
          nullable: true
          items: { $ref: "#/components/schemas/MeetingActionItem" }
        personId:        { type: string, nullable: true }
        organizationId:        { type: string, nullable: true }
        householdId:     { type: string, nullable: true }
    PromoteActionItemBody:
      type: object
      required: [index]
      properties:
        index:           { type: integer, description: "Zero-based index into actionItems[]." }
        assigneeUserId:  { type: string, description: "Optional override; the extracted assigneeName is free-text and may not map to a user." }
        dueDate:         { type: string, format: date }

    # ─── Saved views ───────────────────────────────────────────────────────────
...